Add retry strategy when the leader has changed
It retries write queries when the selected connection return a not leader error.

Retry strategy:
Every 200 milliseconds and expires after 1 second, in order to not return an error to the first query that discovers the leader has changed.

Example:
```bash
iex(1)> Bolt.RoutingDriver.write_query("CREATE (a:Person {name:'Bob'})")

17:28:43.338 [debug] [Bolt.RoutingDriver] core3.docker:7689 query...

17:28:43.338 [debug] CREATE (a:Person {name:'Bob'})

17:28:43.355 [info]  [Bolt.RoutingDriver] Role error notified, refreshing table...

17:28:43.559 [debug] [Bolt.RoutingDriver] core1.docker:7687 query...

17:28:43.559 [debug] CREATE (a:Person {name:'Bob'})
{:ok,
 %{
   stats: %{"labels-added" => 1, "nodes-created" => 1, "properties-set" => 1},
   type: "w"
 }}
```